# backend/api/swagger.yaml
openapi: "3.0.3"
info:
  title: 算力撮合交易平台 API
  description: |
    ## 概述
    算力现货挂牌撮合交易平台 v1 API。包含账户认证、算力撮合、支付分账、居间金融、运营管理、区块链存证六大模块。

    **Base URL**: `http://localhost:8080/api/v1`

    ## 鉴权
    除公开接口外，所有接口需 `Authorization: Bearer <jwt_token>`

    ## 金额单位
    所有金额单位为**分（fen）**，即 `100 = ¥1.00`

    ## 通用响应格式
    ```json
    { "code": 0, "message": "success", "data": {}, "request_id": "xxx" }
    ```
  version: "1.0.0"
  contact:
    name: Platform Team
servers:
  - url: http://localhost:8080/api/v1
    description: 本地开发环境
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - BearerAuth: []
paths:
...
  /user/kyc/personal:
    post:
      tags: [用户 User]
      summary: 个人实名认证（试运行阶段自动通过）
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [real_name, id_card]
              properties:
                real_name: { type: string, example: "张三" }
                id_card: { type: string, example: "110101199001011234" }
  /user/kyc/enterprise:
    post:
      tags: [用户 User]
      summary: 企业认证（试运行阶段自动通过）
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [enterprise_name, uscc, legal_person, legal_person_id_card, bank_name, bank_account_name, bank_account_number, business_license]
              properties:
                enterprise_name: { type: string, example: "某科技有限公司" }
                uscc: { type: string }
                legal_person: { type: string }
                legal_person_id_card: { type: string }
                bank_name: { type: string }
                bank_account_name: { type: string }
                bank_account_number: { type: string }
                business_license: { type: string, format: binary, description: "PDF/JPG/PNG，最大 5MB" }
...
  /supplier/settlements:
    get:
      tags: [支付分账 Payment]
      summary: 供给方结算流水
      description: "按商品归属过滤(settlements.payee_id 不可靠); status: pending/processing/success/failed。"
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [pending, processing, success, failed] } }
        - { name: page, in: query, schema: { type: integer } }
        - { name: page_size, in: query, schema: { type: integer } }
...
  /payment/pay:
    post:
      tags: [支付分账 Payment]
      summary: 发起支付
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [order_no, channel]
              properties:
                order_no: { type: string }
                channel: { type: string, enum: [wechat, alipay, bank] }
  /payment/status/{order_no}:
    get:
      tags: [支付分账 Payment]
      summary: 支付状态查询
      parameters:
        - { name: order_no, in: path, required: true, schema: { type: string } }
  /payment/callback:
    post:
      tags: [支付分账 Payment]
      summary: 易宝支付回调
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                tx_id: { type: string }
                order_no: { type: string }
                status: { type: string }
                amount: { type: integer }
...
  /leads:
    post:
      tags: [居间金融 Intermediary]
      summary: 创建线索
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                type: { type: string, enum: [equipment, construction] }
                contact_name: { type: string }
                contact_phone: { type: string }
                contact_email: { type: string }
                description: { type: string }
                amount_range: { type: string }
                term: { type: string }
...
  /leads/{id}/close:
    post:
      tags: [居间金融 Intermediary]
      summary: 成交登记
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [deal_amount, commission_rate]
              properties:
                deal_amount: { type: integer, description: "成交金额(分)" }
                commission_rate: { type: number, description: "佣金率(%)" }

  # ==================== 运营管理 ====================
...